// src/models/IDocs.model.ts
import { workSpaceInterface } from '../interfaces/workSpaceInterface';

export class Docs implements workSpaceInterface {
    docs: any[];
    workspaceCategory: string;
    userId: string;
    start: string;
    end: string;
    serviceId: string
    duration: string;
    prepaid: string;
    pass: string;
    passType: string;
    limit: number;
    approved: string;
    bill: string;
    indexOf: any;
    splice: any;

}